Cleanup sprite pipeline
